python/requirements.txt
pyjson5
sentence-transformers
python-dotenv
